ピカチュウの画像用のカスタムHook
